Alter window size, block gen, fix collision bug
Prior to this change, the screen size was 600x480, which is too small
for the game. The blocks were too large with too much space in between
them. Sometimes the ball would collide with multiple blocks, causing a
destruction cascade, removing large portions of the blocks at a time.

This change changes the resolution to 800x480. The blocks are now
thinner and have less space in between them, while also leaving space
above them to make the game more fun to play. The collision loop now
breaks upon destroying a single block, which prevents the destruction
cascade, ensuring that the ball will destroy only one block at a time.