Centralize collision detection
Prior to this change, collisions between different types of objects were
handled separately in every object, making it confusing to alter
collision which affects multiple objects.

This change puts all collision detection in the main game loop. It calls
corresponding movement functions in the objects - so we do not need to
expose movement logic.