Create lives counter, alter ball floor collision
Prior to this change, there was no way to lose the ball. Instead, it
would bounce regularly off the floor.

This change creates a ball life counter at the top of the screen. When
the ball hits the floor, it resets to its original position and the game
pauses until player clicks again