Move ballHitPaddle boolean to main from ball
Prior to this change, the ball would keep track of whether it hit the
paddle or not. However, this meant sound effects relied on the general
collision and not the fixed collision.

This change moves the ball-paddle collision logic fix into the main
class. This makes the fix a lot more flexible.